Changed landing page design
