+++ lib/pages/LandingPage/landing_page.dart
@@ -8,6 +8,7 @@ import 'package:instagram_clone/extensions/email_validator.dart';
 import 'package:instagram_clone/pages/Login/user_login_page.dart';
 
 import '../UserRegistration/user_registration_confirmation_email.dart';
+import '../UserRegistration/user_signup.dart';
 
 void main() {
   runApp(LandingPage());
@@ -27,12 +28,10 @@ class LandingPage extends StatelessWidget {
   Future<void> signInWithTwitter() async {
     TwitterAuthProvider twitterProvider = TwitterAuthProvider();
 
-      await FirebaseAuth.instance.signInWithProvider(twitterProvider);
-
+    await FirebaseAuth.instance.signInWithProvider(twitterProvider);
   }
 
   Future<UserCredential> signInWithGoogle() async {
-
     final GoogleSignInAccount? user = await GoogleSignIn().signIn();
 
     final GoogleSignInAuthentication? auth = await user?.authentication;
@@ -45,219 +44,184 @@ class LandingPage extends StatelessWidget {
     return await FirebaseAuth.instance.signInWithCredential(credential);
   }
 
-  Future<void> sendVerificationEmail(BuildContext context) async {
-    final user = FirebaseAuth.instance.currentUser!;
-    await user.sendEmailVerification();
-    if (context.mounted) {
-      Navigator.push(
-        context,
-        MaterialPageRoute(
-          builder: (context) => const UserRegistrationConfirmationEmail(),
-        ),
-      );
-    }
-  }
-
-  Future<void> signUpUser(BuildContext context) async {
-    await FirebaseAuth.instance.createUserWithEmailAndPassword(
-      email: emailController.text.trim(),
-      password: passwordController.text.trim(),
-    );
-
-
-    if (FirebaseAuth.instance.currentUser != null) {
-
-      FirebaseAuth.instance.currentUser!.updateDisplayName(userNameController.text.trim());
-
-      if (context.mounted) {
-        sendVerificationEmail(context);
-      }
-    }
-
-  }
-
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        body: Container(
-          padding: const EdgeInsets.all(40.0),
-          child: SingleChildScrollView(
-            physics: const NeverScrollableScrollPhysics(),
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                const SizedBox(height: 80.0),
-                Image.asset(
-                  'assets/logo.png',
-                  height: 80.0,
-                ),
-                const SizedBox(height: 40.0),
-                ///Needs Verification
-                TextFormField(
-                  key: emailKey,
-                  controller: emailController,
-                  autovalidateMode: AutovalidateMode.onUserInteraction,
-                  decoration: InputDecoration(
-                    labelText: 'Email',
-                    focusedBorder: OutlineInputBorder(
-                      borderSide: BorderSide(
-                        color: appSecondaryColour,
-                      )
-                    ),
-                    focusedErrorBorder: const OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: Colors.red,
-                        )
-                    ),
-                  ),
-                  validator: (value) {
-                    if (!value!.isValidEmail() && value.isNotEmpty) {
-                      return "Invalid Email";
-                    }
-                    return null;
-                  },
-                ),
-                const SizedBox(height: 20.0),
-
-                TextFormField(
-                  key: userNameKey,
-                  controller: userNameController,
-                  decoration: InputDecoration(
-                    labelText: 'Username',
-                    focusedBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: appSecondaryColour,
-                        )
-                    ),
-                    focusedErrorBorder: const OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: Colors.red,
-                        )
-                    ),
-                  ),
-                ),
-                const SizedBox(height: 20.0),
-
-                TextFormField(
-                  key: passwordKey,
-                  controller: passwordController,
-                  obscureText: true,
-                  decoration: InputDecoration(
-                    labelText: 'Password',
-                    focusedBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: appSecondaryColour,
-                        )
-                    ),
-                    focusedErrorBorder: const OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: Colors.red,
-                        )
-                    ),
-                  ),
-                ),
-                const SizedBox(height: 30.0),
-                ElevatedButton(
-                  onPressed: () => signUpUser(context),
-                  child: const Text('Signup'),
-                ),
-                TextButton(
-                    onPressed: () => Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (context) => const LoginPage(
-
-                        ),
-                      ),
-                    ),
-                    child: const Text(
-                      "Login"
-                    ),
-                ),
-                Row(
+      body: Container(
+        padding: const EdgeInsets.all(40.0),
+        child: Column(
+          children: [
+            Padding(
+              padding: const EdgeInsets.only(bottom: 20.0),
+              child: SingleChildScrollView(
+                physics: const NeverScrollableScrollPhysics(),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
                   children: [
+                    const SizedBox(height: 80.0),
+                    Image.asset(
+                      'assets/logo.png',
+                      height: 80.0,
+                    ),
+                    const SizedBox(height: 40.0),
 
-                    const Spacer(),
+                    Center(
+                      child: Column(
+                        children: [
 
-                    Expanded(
-                      child: FractionallySizedBox(
-                        widthFactor: 3,
-                        child: Container(
-                          decoration: BoxDecoration(
-                            border: Border(
-                              bottom: BorderSide(
-                                color: appTertiaryColour,
-                                width: 1
-                              )
-                            )
+                          Row(
+                            children: [
+                              const Spacer(),
+                              Expanded(
+                                child: FractionallySizedBox(
+                                  widthFactor: 3,
+                                  child: Container(
+                                    decoration: BoxDecoration(
+                                        border: Border(
+                                            bottom: BorderSide(
+                                                color: appTertiaryColour, width: 1))),
+                                  ),
+                                ),
+                              ),
+                              const Spacer(),
+                              Container(
+                                color: appPrimaryColour,
+                                padding: const EdgeInsets.only(
+                                  left: 8.0,
+                                  right: 8.0,
+                                ),
+                                margin: const EdgeInsets.all(8.0),
+                                child: Text(
+                                  "SIGN IN",
+                                  style: defaultTextStyle.copyWith(
+                                    fontWeight: FontWeight.w300,
+                                  ),
+                                ),
+                              ),
+                              const Spacer(),
+                              Expanded(
+                                child: FractionallySizedBox(
+                                  widthFactor: 3,
+                                  child: Container(
+                                    decoration: BoxDecoration(
+                                        border: Border(
+                                            bottom: BorderSide(
+                                                color: appTertiaryColour, width: 1))),
+                                  ),
+                                ),
+                              ),
+                              const Spacer(),
+                            ],
                           ),
-                        ),
-                      ),
-                    ),
 
-                    const Spacer(),
+                          const SizedBox(height: 10.0),
 
-                    Container(
-                      color: appPrimaryColour,
-                      padding: const EdgeInsets.only(
-                          left:8.0,
-                          right:8.0,
-                      ),
-                      margin: const EdgeInsets.all(8.0),
-                      child: Text(
-                        "OR",
-                        style: defaultTextStyle.copyWith(
-                          fontWeight: FontWeight.w300,
-                        ),
-                      ),
-                    ),
-
-                    const Spacer(),
-
-                    Expanded(
-                      child: FractionallySizedBox(
-                        widthFactor: 3,
-                        child: Container(
-                          decoration: BoxDecoration(
-                              border: Border(
-                                  bottom: BorderSide(
-                                      color: appTertiaryColour,
-                                      width: 1
-                                  )
-                              )
+                          SignInButton(
+                            Buttons.Email,
+                            text: "Sign in with Email/Phone",
+                            onPressed: () => Navigator.push(
+                              context,
+                              MaterialPageRoute(
+                                builder: (context) => const LoginPage(),
+                              ),
+                            ),
                           ),
-                        ),
+                        ],
                       ),
                     ),
 
-                    const Spacer(),
-
+                    SignInButton(
+                      Buttons.Google,
+                      text: "Sign in with Google",
+                      onPressed: () => signInWithGoogle(),
+                    ),
+                    SignInButton(
+                      Buttons.Twitter,
+                      text: "Sign in with Twitter",
+                      onPressed: () => signInWithTwitter(),
+                    ),
                   ],
                 ),
-
-                //https://developers.google.com/identity/branding-guidelines Branding guidelines
-
-                Center(
-                  child: Column(
-                    children: [
-                      SignInButton(
-                        Buttons.Google,
-                        text: "Sign in with Google",
-                        onPressed: () => signInWithGoogle(),
-                      ),
-                      SignInButton(
-                        Buttons.Twitter,
-                        text: "Sign in with Twitter",
-                        onPressed: () => signInWithTwitter(),
-                      ),
-                    ],
+              ),
+            ),
+            Row(
+              children: [
+                const Spacer(),
+                Expanded(
+                  child: FractionallySizedBox(
+                    widthFactor: 3,
+                    child: Container(
+                      decoration: BoxDecoration(
+                          border: Border(
+                              bottom: BorderSide(
+                                  color: appTertiaryColour, width: 1))),
+                    ),
                   ),
                 ),
+                const Spacer(),
+                Container(
+                  color: appPrimaryColour,
+                  padding: const EdgeInsets.only(
+                    left: 8.0,
+                    right: 8.0,
+                  ),
+                  margin: const EdgeInsets.all(8.0),
+                  child: Text(
+                    "SIGN UP",
+                    style: defaultTextStyle.copyWith(
+                      fontWeight: FontWeight.w300,
+                    ),
+                  ),
+                ),
+                const Spacer(),
+                Expanded(
+                  child: FractionallySizedBox(
+                    widthFactor: 3,
+                    child: Container(
+                      decoration: BoxDecoration(
+                          border: Border(
+                              bottom: BorderSide(
+                                  color: appTertiaryColour, width: 1))),
+                    ),
+                  ),
+                ),
+                const Spacer(),
               ],
             ),
-          ),
+            Padding( // This stays at the bottom
+              padding: const EdgeInsets.only(
+                top: 20.0,
+                bottom: 20.0,
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: <Widget>[
+                  const Text('Don\'t have an account?'),
+                  const SizedBox(width: 4.0),
+                  GestureDetector(
+                    onTap: () {
+                      Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                          builder: (context) => const SignupPage(),
+                        ),
+                      );
+                    },
+                    child: const Text(
+                      'Sign Up',
+                      style: TextStyle(
+                        color: Colors.blue,
+                        fontWeight: FontWeight.w600,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            const Spacer(flex: 1),
+          ],
         ),
+      ),
     );
   }
 }
