--- lib/pages/UserRegistration/user_registration_page.dart
@@ -1,133 +1,0 @@
-import 'package:flutter/material.dart';
-import 'user_registration_confirmation_page.dart';
-import 'user_registration_email_signup.dart';
-
-class SignupPage extends StatefulWidget {
-  const SignupPage({Key? key}) : super(key: key);
-
-  @override
-  SignupPageState createState() => SignupPageState();
-}
-
-class SignupPageState extends State<SignupPage> {
-  final _phoneNumberController = TextEditingController();
-
-  @override
-  void dispose() {
-    _phoneNumberController.dispose();
-    super.dispose();
-  }
-
-  void _navigateToEnterConfirmationPage(BuildContext context) {
-    String phoneNumber = _phoneNumberController.text;
-    if (phoneNumber.isEmpty) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text('Please enter a phone number.'),
-        ),
-      );
-    } else {
-      Navigator.pushReplacement(
-        context,
-        MaterialPageRoute(
-          builder: (context) => const EnterConfirmationPage(),
-        ),
-      );
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        leading: IconButton(
-          icon: const Icon(Icons.arrow_back),
-          onPressed: () {
-            Navigator.pop(context);
-          },
-        ),
-        backgroundColor: Colors.blue,
-      ),
-      body: SingleChildScrollView( // Added SingleChildScrollView
-        child: Container(
-          alignment: Alignment.center,
-          padding: const EdgeInsets.symmetric(horizontal: 24.0),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: <Widget>[
-              Column(
-                children: <Widget>[
-                  const SizedBox(height: 80.0),
-                  Image.asset(
-                    'assets/instaclone.png',
-                    height: 80.0,
-                  ),
-                  const SizedBox(height: 40.0),
-                  const Text(
-                    "What's your mobile number?",
-                    style: TextStyle(
-                      fontSize: 24.0,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                  const SizedBox(height: 30.0),
-                  const Text(
-                    "Enter the mobile number on which you can be contacted. No one will see this on your profile.",
-                    style: TextStyle(
-                      fontSize: 16,
-                    ),
-                  ),
-                  const SizedBox(height: 20.0),
-                  TextField(
-                    controller: _phoneNumberController,
-                    keyboardType: TextInputType.phone,
-                    decoration: const InputDecoration(
-                      labelText: "Mobile number",
-                      labelStyle: TextStyle(color: Colors.grey),
-                      border: OutlineInputBorder(),
-                    ),
-                  ),
-                  const SizedBox(height: 8),
-                  const Text(
-                    "You may receive SMS notifications from us for security and login purposes.",
-                    style: TextStyle(
-                      fontSize: 14,
-                      color: Colors.grey,
-                    ),
-                  ),
-                  const SizedBox(height: 30.0),
-                  Center(
-                    child: ElevatedButton(
-                      onPressed: () => _navigateToEnterConfirmationPage(context),
-                      style: ButtonStyle(
-                        backgroundColor: MaterialStateProperty.all(Colors.blue),
-                      ),
-                      child: const Text("Next"),
-                    ),
-                  ),
-                ],
-              ),
-              Center(
-                child: TextButton(
-                  onPressed: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (context) => UserRegistrationEmailSignup(),
-                      ),
-                    );
-                  },
-                  style: ButtonStyle(
-                    foregroundColor: MaterialStateProperty.all(Colors.blue),
-                  ),
-                  child: const Text("Sign up with email address"),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
